+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.0.0] - Unreleased
+### Changed
+- **BREAKING**:`usePath` returns `decodeURIComponent`-ed path
+- **BREAKING**:`useRoutes` matches paths that have been `decodeURIComponent`-ed (e.g. `/weird (route)` will match if provided as a key to `useRoutes`)
+### Fixed
+- `useRoutes` props type to `string`
+
 ## [2.6.0] - 2021-10-15
 ### Added
 - `useMatch` hook